PreAnalyzedField tokenStreamComponents reuse fix
We have to manually call clearAttributes() in decodeInput(...) before the attributes
are removed. Attribute removal is important for reuse, but also means that the
clearAttributes() call in parser.parse(reader, this) has no effect (hence us calling
it in decodeInput(...)). There are other places we could call this, but placing it near
removeAllAttributes() seems to make sense.
NOTE: I don't understand why this appears to have become an issue sometime between
da8ea706826..9805b125dc8.